LayoutTests/media/loadedmetadata-fires-without-user-gesture-when-setRequiresUserGestureToLoadVideo-false.html

<!DOCTYPE html>
<html>
<head>
<script src="media-file.js"></script>
<script src="video-test.js"></script>
<script>
if (window.internals) {
    internals.settings.setVideoPlaybackRequiresUserGesture(true);
    internals.settings.setRequiresUserGestureToLoadVideo(false);
}

function done()
{
    passTest("Did load metadata without user gesture");
}

function start()
{
    findMediaElement();
    waitForEvent("loadedmetadata", done);
    video.src = findMediaFile("video", "content/test");
}
</script>
</head>
<body onload="start()">
    <video id="video"></video>
</body>
</html>
