Safari-611.1.10/LayoutTests/media/loadedmetadata-fires-without-user-gesture-when-setRequiresUserGestureToLoadVideo-false-expected.txt


EVENT(loadedmetadata)
Did load metadata without user gesture OK
END OF TEST
